Compila, no faultea, pero dibuja los triangulos en cualquier lado

+++ copito.h
@@ -7,5 +7,5 @@ float calc_tol (float x1, float y1, float x2, float y2);
 
 //Variables necesarias
 extern uint8_t orden;   //En el main
-extern uint8_t tolerancia; 
+extern float tolerancia; 
 #endif
